Pull out Balanced logic to a BalancedScoreTracker
